Make vsat the default database
When running locally we want `vsat` to be the default database.

Currently the default is not `vsat` and the (local) connection string assumees that. It worked fine (for me) because at some point I created the `vsat` DB manually and yeah... it doesn't work out of the box for anyone else.