Use published state for story image
Currently when we publish a story we persist the state of the published story... cool. If the first scene in the story has an image -- it's optional -- then the URLs to that scene/story image are persisted.

When we're retrieving a published story summary we _don't_ interrogate that state to get the first scene's image (if any); instead, we're joining to the _current_ state of the story which may be using an entirely different image for that opening scene.

This is... not good. When a story is published, it's a snapshot of the story at the time of publication. The "cover" mustn't suddenly change underfoot.

The changes here ensure that the image is static.